+++ src/tests/test.verification.ts
@@ -1,4 +1,4 @@
-import { Proof, responseSelection } from '../types'
+import { Proof } from '../types'
 import { generateUuid } from '../utils'
 import { reclaimprotocol } from '../'
 
@@ -46,102 +46,6 @@ describe('Verification', () => {
 	})
 })
 
-describe('Extract parameters', () => {
-	it('should extract parameters', () => {
-		const responseSelection: responseSelection[] = [
-			{
-				responseMatch: '<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“>{{username}}</span>.*?<span>{{karma}} karma</span>'
-			},
-			{
-				responseMatch: '<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“>{{user}}</span>.*?<span>{{dharma}} karma</span>'
-			}
-		]
-		const proof: Proof = {
-			'onChainClaimId': '2617',
-			'templateClaimId':'0',
-			'provider':'http',
-			'parameters': {
-				'url':'https://www.reddit.com',
-				'method':'GET',
-				'responseSelections':[
-					{ 'responseMatch':'<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“>Hairy-Firefighter</span>.*?<span>10 karma</span>' },
-					{ 'responseMatch':'<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“>Hairy</span>.*?<span>20 karma</span>' }
-				]
-			},
-			'chainId':420,
-			'ownerPublicKey':'03057dd1b36d108cc0d9bced0565b6363ed910bc6522aa937092e1dc344614ddde',
-			'timestampS':'1684343138',
-			'witnessAddresses':['reclaim-node.questbook.app'],
-			'signatures':['0xea2b6a7f7183ddea565a0cb569ce4ff2896f3b6f4dcdd2b5da5be1e67f8865086fcff48227336636a3a32c9cbf801d946351b00b411b2e129ad4b005acfb4fed1b'],
-			'redactedParameters':'{"url":"**********************@undefined"}'
-		}
-		const parametersExtracted = reclaimprotocol.utils.extractParameterValues(responseSelection, proof)
-
-		expect(Object.keys(parametersExtracted).length).toBe(4)
-		// expect(parametersExtracted['username']).toBe('Hairy-Firefighter')
-		// expect(parametersExtracted['karma']).toBe('10')
-
-	})
-
-	it('should return empty parameter object', () => {
-		const responseSelection: responseSelection[] = [
-			{
-				responseMatch: '<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“></span>.*?<span> karma</span>'
-			}
-		]
-		const proof: Proof = {
-			'onChainClaimId': '2617',
-			'templateClaimId':'0',
-			'provider':'http',
-			'parameters': {
-				'url':'https://www.reddit.com',
-				'method':'GET',
-				'responseSelections':[
-					{ 'responseMatch':'<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“>Hairy-Firefighter</span>.*?<span>10 karma</span>' }
-				]
-			},
-			'chainId':420,
-			'ownerPublicKey':'03057dd1b36d108cc0d9bced0565b6363ed910bc6522aa937092e1dc344614ddde',
-			'timestampS':'1684343138',
-			'witnessAddresses':['reclaim-node.questbook.app'],
-			'signatures':['0xea2b6a7f7183ddea565a0cb569ce4ff2896f3b6f4dcdd2b5da5be1e67f8865086fcff48227336636a3a32c9cbf801d946351b00b411b2e129ad4b005acfb4fed1b'],
-			'redactedParameters':'{"url":"**********************@undefined"}'
-		}
-		const parametersExtracted = reclaimprotocol.utils.extractParameterValues(responseSelection, proof)
-		expect(Object.keys(parametersExtracted).length).toBe(0)
-	})
-
-	it('should throw incorrect response selection', () => {
-		const responseSelection: responseSelection[] = [
-			{
-				responseMatch: '<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“>{{username}}</span>.*?<span>{{karma}} karma</span>'
-			},
-			{
-				responseMatch: '<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“>{{username}}</span>.*?<span>{{karma}} karma</span>'
-			}
-		]
-		const proof: Proof = {
-			'onChainClaimId': '2617',
-			'templateClaimId':'0',
-			'provider':'http',
-			'parameters': {
-				'url':'https://www.reddit.com',
-				'method':'GET',
-				'responseSelections':[
-					{ 'responseMatch':'<span class=\“_2BMnTatQ5gjKGK5OWROgaG\“>Hairy-Firefighter</span>.*?<span>10 karma</span>' }
-				]
-			},
-			'chainId':420,
-			'ownerPublicKey':'03057dd1b36d108cc0d9bced0565b6363ed910bc6522aa937092e1dc344614ddde',
-			'timestampS':'1684343138',
-			'witnessAddresses':['reclaim-node.questbook.app'],
-			'signatures':['0xea2b6a7f7183ddea565a0cb569ce4ff2896f3b6f4dcdd2b5da5be1e67f8865086fcff48227336636a3a32c9cbf801d946351b00b411b2e129ad4b005acfb4fed1b'],
-			'redactedParameters':'{"url":"**********************@undefined"}'
-		}
-		expect(() => reclaimprotocol.utils.extractParameterValues(responseSelection, proof)).toThrow('Invalid response selections')
-	})
-})
-
 
 const CORRECT_PROOF: Proof = {
 	onChainClaimId: '1560',
@@ -157,79 +61,31 @@ const CORRECT_PROOF: Proof = {
 }
 
 const INCORRECT_SIGNATURE_PROOF: Proof = {
-	onChainClaimId: '1560',
-	templateClaimId: generateUuid(),
-	chainId: 420,
-	provider: 'google-login',
-	parameters: { 'emailAddress': 'swetasunofficial@gmail.com' },
-	ownerPublicKey: '039549ccde10c559c979eb826075e9274ed8d9439e299e46f752fc8e9cd1e0647f',
-	timestampS: '1681968148',
-	witnessAddresses: ['reclaim-node.questbook.app'],
+	...CORRECT_PROOF,
 	signatures: ['0x86846dc92c08f27c646f87b9e58e3abd20005621c9113ae9c3718a6f71d882511b43b3f373b4e941d7d5636106dfdcf9807496d6b9f1a5be8320fe2a1afebf521c'],
-	redactedParameters: '{"emailAddress":"****************@gmail.com"}'
 }
 
 const INCORRECT_OWNER_PUBLIC_KEY_PROOF: Proof = {
-	onChainClaimId: '1560',
-	templateClaimId: generateUuid(),
-	chainId: 420,
-	provider: 'google-login',
-	parameters: { 'emailAddress': 'swetasunofficial@gmail.com' },
+	...CORRECT_PROOF,
 	ownerPublicKey: '038549ccde10c559c979eb826075e9274ed8d9439e299e46f752fc8e9cd1e0647f',
-	timestampS: '1681968148',
-	witnessAddresses: ['reclaim-node.questbook.app'],
-	signatures: ['0x72846dc92c08f27c646f87b9e58e3abd20005621c9113ae9c3718a6f71d882511b43b3f373b4e941d7d5636106dfdcf9807496d6b9f1a5be8320fe2a1afebf521c'],
-	redactedParameters: '{"emailAddress":"****************@gmail.com"}'
 }
 
 const INCORRECT_TIMESTAMP_PROOF: Proof = {
-	onChainClaimId: '1560',
-	templateClaimId: generateUuid(),
-	chainId: 420,
-	provider: 'google-login',
-	parameters: { 'emailAddress': 'swetasunofficial@gmail.com' },
-	ownerPublicKey: '039549ccde10c559c979eb826075e9274ed8d9439e299e46f752fc8e9cd1e0647f',
+	...CORRECT_PROOF,
 	timestampS: '1681968149',
-	witnessAddresses: ['reclaim-node.questbook.app'],
-	signatures: ['0x72846dc92c08f27c646f87b9e58e3abd20005621c9113ae9c3718a6f71d882511b43b3f373b4e941d7d5636106dfdcf9807496d6b9f1a5be8320fe2a1afebf521c'],
-	redactedParameters: '{"emailAddress":"****************@gmail.com"}'
 }
 
 const INCORRECT_PARAMETER_PROOF: Proof = {
-	onChainClaimId: '1560',
-	templateClaimId: generateUuid(),
-	chainId: 420,
-	provider: 'google-login',
+	...CORRECT_PROOF,
 	parameters: { 'emailAddress': 'sweta@gmail.com' },
-	ownerPublicKey: '039549ccde10c559c979eb826075e9274ed8d9439e299e46f752fc8e9cd1e0647f',
-	timestampS: '1681968148',
-	witnessAddresses: ['reclaim-node.questbook.app'],
-	signatures: ['0x72846dc92c08f27c646f87b9e58e3abd20005621c9113ae9c3718a6f71d882511b43b3f373b4e941d7d5636106dfdcf9807496d6b9f1a5be8320fe2a1afebf521c'],
-	redactedParameters: '{"emailAddress":"****************@gmail.com"}'
 }
 
 const INCORRECT_CLAIM_ID_PROOF: Proof = {
+	...CORRECT_PROOF,
 	onChainClaimId: '1561',
-	templateClaimId: generateUuid(),
-	chainId: 420,
-	provider: 'google-login',
-	parameters: { 'emailAddress': 'swetasunofficial@gmail.com' },
-	ownerPublicKey: '039549ccde10c559c979eb826075e9274ed8d9439e299e46f752fc8e9cd1e0647f',
-	timestampS: '1681968148',
-	witnessAddresses: ['reclaim-node.questbook.app'],
-	signatures: ['0x72846dc92c08f27c646f87b9e58e3abd20005621c9113ae9c3718a6f71d882511b43b3f373b4e941d7d5636106dfdcf9807496d6b9f1a5be8320fe2a1afebf521c'],
-	redactedParameters: '{"emailAddress":"****************@gmail.com"}'
 }
 
 const INCORRECT_PROVIDER_PROOF: Proof = {
-	onChainClaimId: '1560',
-	templateClaimId: generateUuid(),
-	chainId: 420,
+	...CORRECT_PROOF,
 	provider: 'yc-login',
-	parameters: { 'emailAddress': 'swetasunofficial@gmail.com' },
-	ownerPublicKey: '039549ccde10c559c979eb826075e9274ed8d9439e299e46f752fc8e9cd1e0647f',
-	timestampS: '1681968148',
-	witnessAddresses: ['reclaim-node.questbook.app'],
-	signatures: ['0x72846dc92c08f27c646f87b9e58e3abd20005621c9113ae9c3718a6f71d882511b43b3f373b4e941d7d5636106dfdcf9807496d6b9f1a5be8320fe2a1afebf521c'],
-	redactedParameters: '{"emailAddress":"****************@gmail.com"}'
 }
